Move JdbiPlugin to core.  Javadoc fixes.  Nuke DefaultStatementBuilderFactory
