+++ core/src/test/java/org/jdbi/v3/core/TestPlugins.java
@@ -19,7 +19,6 @@ import static org.mockito.Mockito.mock;
 import java.sql.Connection;
 
 import org.jdbi.v3.core.rule.H2DatabaseRule;
-import org.jdbi.v3.core.spi.JdbiPlugin;
 import org.junit.Rule;
 import org.junit.Test;
