+++ oracle12/src/test/java/org/jdbi/v3/oracle12/OracleDatabaseRule.java
@@ -24,8 +24,8 @@ import java.util.List;
 import org.jdbi.v3.core.ConnectionFactory;
 import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.Jdbi;
+import org.jdbi.v3.core.JdbiPlugin;
 import org.jdbi.v3.core.rule.DatabaseRule;
-import org.jdbi.v3.core.spi.JdbiPlugin;
 import org.junit.rules.ExternalResource;
 
 /**
