+++ sqlobject/src/test/java/org/jdbi/v3/sqlobject/TestOnDemandSqlObject.java
@@ -32,11 +32,11 @@ import org.jdbi.v3.core.CloseException;
 import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.Jdbi;
 import org.jdbi.v3.core.JdbiException;
+import org.jdbi.v3.core.JdbiPlugin;
 import org.jdbi.v3.core.Something;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.mapper.SomethingMapper;
 import org.jdbi.v3.core.result.ResultIterator;
-import org.jdbi.v3.core.spi.JdbiPlugin;
 import org.jdbi.v3.core.statement.StatementContext;
 import org.jdbi.v3.core.transaction.TransactionException;
 import org.jdbi.v3.sqlobject.customizer.Bind;
